+++ spinta/backends/__init__.py
@@ -32,7 +32,8 @@ from spinta.commands import gen_object_id
 from spinta.commands import is_object_id
 from spinta.commands import load_operator_value
 from spinta.commands import prepare
-from spinta.components import Action, UrlParams, page_in_data
+from spinta.components import UrlParams, page_in_data
+from spinta.core.enums import Action
 from spinta.components import Context
 from spinta.components import DataItem
 from spinta.components import Model