+++ spinta/core/enums.py
@@ -1,4 +1,8 @@
 import enum
+from typing import Union, Any
+
+from spinta import exceptions
+from spinta.utils.errors import report_error
 
 
 class Access(enum.IntEnum):
@@ -39,3 +43,62 @@ class Level(enum.IntEnum):
 
     # Data is linked with a known vocabulary.
     linked = 5
+
+
+class Action(enum.Enum):
+    INSERT = 'insert'
+    UPSERT = 'upsert'
+    UPDATE = 'update'
+    PATCH = 'patch'
+    DELETE = 'delete'
+
+    WIPE = 'wipe'
+
+    GETONE = 'getone'
+    GETALL = 'getall'
+    SEARCH = 'search'
+
+    CHANGES = 'changes'
+
+    CHECK = 'check'
+    INSPECT = 'inspect'
+    SCHEMA = 'schema'
+
+    @classmethod
+    def has_value(cls, value):
+        return value in cls._value2member_map_
+
+    @classmethod
+    def by_value(cls, value):
+        return cls._value2member_map_[value]
+
+    @classmethod
+    def values(cls):
+        return list(cls._value2member_map_.keys())
+
+
+class Mode(enum.Enum):
+    # Internal mode always use internal backend set on manifest, namespace or
+    # model.
+    internal = 'internal'
+
+    # External model always sue external backend set on dataset or model's
+    # source entity.
+    external = 'external'
+
+
+def action_from_op(
+    scope: Any,
+    payload: dict,
+    stop_on_error: bool = True,
+) -> Union[Action, exceptions.UserError]:
+    action = payload.get('_op')
+    if not Action.has_value(action):
+        error = exceptions.UnknownAction(
+            scope,
+            action=action,
+            supported_actions=Action.values(),
+        )
+        report_error(error, stop_on_error=stop_on_error)
+        return error
+    return Action.by_value(action)