+++ tests/formats/test_helpers.py
@@ -8,7 +8,7 @@ from spinta.auth import AdminToken
 from spinta.formats.helpers import get_model_tabular_header
 from spinta.testing.manifest import load_manifest_and_context
 from spinta.testing.request import make_get_request
-from spinta.components import Action
+from spinta.core.enums import Action
 from spinta.components import UrlParams
 from spinta.components import Version
 from spinta.core.config import RawConfig