+++ tests/formats/test_html.py
@@ -14,7 +14,7 @@ from starlette.requests import Request
 from starlette.datastructures import Headers
 
 from spinta import commands
-from spinta.components import Action
+from spinta.core.enums import Action
 from spinta.components import Config
 from spinta.components import Context
 from spinta.components import Namespace